badge23: scope: better locking
This uses atomics to lock out concurrenct writes/reads, and fixes some
potential null pointer dereferences.

+++ components/badge23/include/badge23/scope.h
@@ -5,7 +5,8 @@ typedef struct {
     int16_t * buffer;
     int16_t buffer_size;
     int16_t write_head_position; // last written value
-    volatile uint8_t is_being_read;
+    // atomic value used to prevent simultaneous reads/writes (not a mutex!).
+    volatile uint32_t is_being_read;
 } scope_t;
 
 void init_scope(uint16_t size);
